NewsService 1.2

+++ src/main/java/com/example/NewsService/repository/CommentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    List<Comment> findByNewsId(Long newsId);
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+    List<Comment> findByNewsId(int newsId);
 }
